+++ src/renderer/store/schema/ids.js
@@ -1,19 +1,24 @@
 import * as L from '../../../shared/level'
-import { REDUNDANT_IDENTIFIERS } from './schema'
 
-export default async ({ schemaDB, jsonDB, options }) => {
-  const actual = await L.get(schemaDB, REDUNDANT_IDENTIFIERS, true)
-  const wanted = options[REDUNDANT_IDENTIFIERS]
+const upgrade = async jsonDB => {
+  const tuples = await L.readTuples(jsonDB, '')
+  const ops = tuples
+    .filter(([_, value]) => value.id)
+    .map(([key, { id, ...value }]) => L.putOp(key, value))
 
-  const upgrade = async () => {
-    const tuples = await L.readTuples(jsonDB, '')
-    const ops = tuples
-      .filter(([_, value]) => value.id)
-      .map(([key, { id, ...value }]) => L.putOp(key, value))
+  await jsonDB.batch(ops)
+}
 
-    await this.jsonDB.batch(ops)
-    await this.schemaDB.put(REDUNDANT_IDENTIFIERS, false)
-  }
+const downgrade = async jsonDB => {
+  // TODO: limit to scopes originally carrying ids in values
+  const tuples = await L.readTuples(jsonDB, '')
+  const ops = tuples
+    .map(([key, value]) => L.putOp(key, { ...value, id: key }))
 
-  if (actual && wanted === false) await upgrade()
+  await jsonDB.batch(ops)
+}
+
+export default {
+  'KEY-ONLY': upgrade,
+  VALUE: downgrade
 }
